chore(sentry): move functions from sentry config to sentry utils
